modify common data ctime, mtime data_type

+++ src/main/java/com/example/wechatmerchant/repository/db/entity/CommonEntity.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class CommonEntity {
 
-    private Integer ctime;
+    private Long ctime;
 
-    private Integer mtime;
+    private Long mtime;
 }
